Testing authOptions export error fix
