+++ src/app/api/glucose/[id]/route.js
@@ -2,7 +2,7 @@ import { NextResponse } from "next/server";
 import { connectMongoDB } from "../../../lib/mongodb";
 import Glucose from "../../../models/glucose";
 import { getServerSession } from "next-auth";
-import { authOptions } from "../../auth/[...nextauth]/route";
+import { authOptions } from "../../auth/[...nextauth]/authOptions";
 
 // Handles the PUT request to edit glucose readings
 export async function PUT(request, { params }) {
