+++ src/app/api/glucose/route.js
@@ -2,7 +2,7 @@ import { NextResponse } from "next/server";
 import { connectMongoDB } from "../../lib/mongodb";
 import Glucose from "../../models/glucose";
 import { getServerSession } from "next-auth";
-import { authOptions } from "../auth/[...nextauth]/route";
+import { authOptions } from "../../auth/[...nextauth]/authOptions";
 
 // Handles the GET request to fetch glucose readings
 export async function GET(request) {
